Renamed test classes
